+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/cart/CartFreezeCartActionImpl.java
@@ -5,6 +5,7 @@ import java.time.*;
 import java.util.*;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.*;
 
 import io.vrap.rmf.base.client.ModelBase;
@@ -16,18 +17,35 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 
 /**
- *  <p>Changes the <a href="https://docs.commercetools.com/apis/ctp:api:type:CartState" rel="nofollow">CartState</a> from <code>Active</code> to <code>Frozen</code>. Results in a <a href="https://docs.commercetools.com/apis/ctp:api:type:FrozenCarts" rel="nofollow">Frozen Cart</a>. Fails with <a href="https://docs.commercetools.com/apis/ctp:api:type:InvalidOperationError" rel="nofollow">InvalidOperation</a> error when the Cart is empty.</p>
- *  <p>Freezing a Cart produces the <a href="https://docs.commercetools.com/apis/ctp:api:type:CartFrozenMessage" rel="nofollow">CartFrozen</a> Message.</p>
+ *  <p>Freezes the Cart based on the provided <a href="https://docs.commercetools.com/apis/ctp:api:type:FreezeStrategy" rel="nofollow">FreezeStrategy</a>.</p>
+ *  <p>The following behavior occurs:</p>
+ *  <ul>
+ *   <li>Changes the Cart State from <code>Active</code> to <code>Frozen</code>.</li>
+ *   <li>Sets the corresponding <a href="https://docs.commercetools.com/apis/ctp:api:type:FreezeStrategy" rel="nofollow">FreezeStrategy</a> on the Cart's <code>freezeStrategy</code> field.</li>
+ *   <li>Produces the <a href="https://docs.commercetools.com/apis/ctp:api:type:CartFrozenMessage" rel="nofollow">CartFrozen</a> Message.</li>
+ *  </ul>
+ *  <p>If the Cart is empty, an <a href="https://docs.commercetools.com/apis/ctp:api:type:InvalidOperationError" rel="nofollow">InvalidOperation</a> error is returned.</p>
  */
 @Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
 public class CartFreezeCartActionImpl implements CartFreezeCartAction, ModelBase {
 
     private String action;
 
+    private com.commercetools.api.models.cart.FreezeStrategy strategy;
+
     /**
      * create instance with all properties
      */
     @JsonCreator
+    CartFreezeCartActionImpl(
+            @JsonProperty("strategy") final com.commercetools.api.models.cart.FreezeStrategy strategy) {
+        this.strategy = strategy;
+        this.action = FREEZE_CART;
+    }
+
+    /**
+     * create empty instance
+     */
     public CartFreezeCartActionImpl() {
         this.action = FREEZE_CART;
     }
@@ -40,6 +58,18 @@ public class CartFreezeCartActionImpl implements CartFreezeCartAction, ModelBase
         return this.action;
     }
 
+    /**
+     *  <p>Strategy that determines the freezing behavior.</p>
+     */
+
+    public com.commercetools.api.models.cart.FreezeStrategy getStrategy() {
+        return this.strategy;
+    }
+
+    public void setStrategy(final com.commercetools.api.models.cart.FreezeStrategy strategy) {
+        this.strategy = strategy;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o)
@@ -50,17 +80,23 @@ public class CartFreezeCartActionImpl implements CartFreezeCartAction, ModelBase
 
         CartFreezeCartActionImpl that = (CartFreezeCartActionImpl) o;
 
-        return new EqualsBuilder().append(action, that.action).append(action, that.action).isEquals();
+        return new EqualsBuilder().append(action, that.action)
+                .append(strategy, that.strategy)
+                .append(action, that.action)
+                .append(strategy, that.strategy)
+                .isEquals();
     }
 
     @Override
     public int hashCode() {
-        return new HashCodeBuilder(17, 37).append(action).toHashCode();
+        return new HashCodeBuilder(17, 37).append(action).append(strategy).toHashCode();
     }
 
     @Override
     public String toString() {
-        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE).append("action", action).build();
+        return new ToStringBuilder(this, ToStringStyle.SHORT_PREFIX_STYLE).append("action", action)
+                .append("strategy", strategy)
+                .build();
     }
 
     @Override